CALCULATOR/requirements.txt
flet==0.27.6
python_dateutil==2.9.0.post0
